📝 (LICENSE, README.md): add MIT license to the project
The MIT license is a permissive license that allows for the free use, modification, and distribution of software.

Adding a license to the project makes it clear to others what they are allowed to do with the code and helps to ensure that the project can be widely used and adopted.